+++ rust-engine/src/backends/lean.rs
@@ -10,6 +10,7 @@ use std::sync::LazyLock;
 use super::prelude::*;
 use crate::{
     ast::identifiers::global_id::view::{ConstructorKind, PathSegment, TypeDefKind},
+    attributes::hax_attributes,
     phase::*,
 };
 
@@ -22,6 +23,8 @@ mod binops {
 use crate::names::rust_primitives::hax::explicit_monadic::{lift, pure};
 const LIFT: GlobalId = lift;
 const PURE: GlobalId = pure;
+
+use hax_lib_macros_types::AttrPayload;
 
 /// The Lean printer
 #[setup_printer_struct]
@@ -470,6 +473,15 @@ const _: () = {
             if spec.precondition.is_none() && spec.postcondition.is_none() {
                 nil!()
             } else {
+                let proofs: Vec<&String> = hax_attributes(&item.meta.attributes)
+                    .flat_map(|attr| match attr {
+                        AttrPayload::Proof(proof) => Some(proof),
+                        _ => None,
+                    })
+                    .collect();
+                if proofs.len() > 1 {
+                    emit_error!("Only one proof attribute per item is allowed.");
+                }
                 docs![
                     hardline!(),
                     hardline!(),
@@ -542,9 +554,17 @@ const _: () = {
                         hardline!(),
                         "pureEnsures := by constructor; intros; mvcgen <;> try grind",
                         hardline!(),
-                        docs!["contract := by mvcgen[", name, "] <;> try grind"]
-                            .group()
-                            .nest(INDENT),
+                        docs![
+                            "contract :=",
+                            line!(),
+                            if proofs.is_empty() {
+                                docs!["by mvcgen[", name, "] <;> try grind"]
+                            } else {
+                                docs![intersperse!(proofs, nil!())]
+                            }
+                        ]
+                        .group()
+                        .nest(INDENT),
                         hardline!(),
                     ]
                     .nest(INDENT)