+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Changes to the Lean backend:
    extra parameters (#1803)
  - Add generation of specs from requires/ensures-annotations (#1815)
  - Add support for nonliteral array sizes (#1826)
+ - Add `hax_lib::lean::proof` attribute (#1831)
 
 Miscellaneous:
 - Reserve extraction folder for auto-generated files in Lean examples (#1754)